+++ src/flags/sorting.rs
@@ -37,6 +37,7 @@ impl Sorting {
 #[derive(Clone, Debug, Copy, PartialEq, Eq, Deserialize)]
 #[serde(rename_all = "kebab-case")]
 pub enum SortColumn {
+    None,
     Extension,
     Name,
     Time,
@@ -62,6 +63,8 @@ impl Configurable<Self> for SortColumn {
             Some(Self::Extension)
         } else if matches.is_present("versionsort") || sort == Some("version") {
             Some(Self::Version)
+        } else if matches.is_present("no-sort") || sort == Some("none") {
+            Some(Self::None)
         } else {
             None
         }
@@ -171,6 +174,10 @@ impl Configurable<Self> for DirGrouping {
             return Some(Self::None);
         }
 
+        if matches.is_present("group-directories-first") {
+            return Some(Self::First);
+        }
+
         if matches.occurrences_of("group-dirs") > 0 {
             if let Some(group_dirs) = matches.values_of("group-dirs")?.last() {
                 return Self::from_str(group_dirs);
@@ -260,6 +267,16 @@ mod test_sort_column {
     }
 
     #[test]
+    fn test_from_arg_matches_no_sort() {
+        let argv = vec!["lsd", "--no-sort"];
+        let matches = app::build().get_matches_from_safe(argv).unwrap();
+        assert_eq!(
+            Some(SortColumn::None),
+            SortColumn::from_arg_matches(&matches)
+        );
+    }
+
+    #[test]
     fn test_from_arg_matches_sort() {
         let argv = vec!["lsd", "--sort", "time"];
         let matches = app::build().get_matches_from_safe(argv).unwrap();
@@ -286,6 +303,13 @@ mod test_sort_column {
         let matches = app::build().get_matches_from_safe(argv).unwrap();
         assert_eq!(
             Some(SortColumn::Version),
+            SortColumn::from_arg_matches(&matches)
+        );
+
+        let argv = vec!["lsd", "--sort", "none"];
+        let matches = app::build().get_matches_from_safe(argv).unwrap();
+        assert_eq!(
+            Some(SortColumn::None),
             SortColumn::from_arg_matches(&matches)
         );
     }
@@ -524,6 +548,16 @@ mod test_dir_grouping {
         let matches = app::build().get_matches_from_safe(argv).unwrap();
         assert_eq!(
             Some(DirGrouping::Last),
+            DirGrouping::from_arg_matches(&matches)
+        );
+    }
+
+    #[test]
+    fn test_from_arg_matches_group_directories_first() {
+        let argv = vec!["lsd", "--group-directories-first"];
+        let matches = app::build().get_matches_from_safe(argv).unwrap();
+        assert_eq!(
+            Some(DirGrouping::First),
             DirGrouping::from_arg_matches(&matches)
         );
     }