+++ src/meta/date.rs
@@ -56,7 +56,7 @@ impl Date {
             match &flags.date {
                 DateFlag::Date => val.format("%c").to_string(),
                 DateFlag::Relative => format!("{}", HumanTime::from(*val - Local::now())),
-                DateFlag::ISO => {
+                DateFlag::Iso => {
                     // 365.2425 * 24 * 60 * 60 = 31556952 seconds per year
                     // 15778476 seconds are 6 months
                     if *val > Local::now() - Duration::seconds(15_778_476) {
@@ -83,7 +83,6 @@ mod test {
     use std::io;
     use std::path::Path;
     use std::process::{Command, ExitStatus};
-    use std::time;
     use std::{env, fs};
 
     #[cfg(unix)]
@@ -267,7 +266,7 @@ mod test {
         let date = Date::from(&file_path.metadata().unwrap());
 
         let mut flags = Flags::default();
-        flags.date = DateFlag::ISO;
+        flags.date = DateFlag::Iso;
 
         assert_eq!(
             creation_date
@@ -295,7 +294,7 @@ mod test {
         let date = Date::from(&file_path.metadata().unwrap());
 
         let mut flags = Flags::default();
-        flags.date = DateFlag::ISO;
+        flags.date = DateFlag::Iso;
 
         assert_eq!(
             creation_date
@@ -314,8 +313,8 @@ mod test {
         // 4437052 is the bad year taken from https://github.com/Peltoche/lsd/issues/529 that we know is both
         // a) high enough to break chrono
         // b) not high enough to break SystemTime (as Duration::MAX would)
-        let end_time =
-            time::SystemTime::UNIX_EPOCH + time::Duration::new(4437052 * 365 * 24 * 60 * 60, 0);
+        let end_time = std::time::SystemTime::UNIX_EPOCH
+            + std::time::Duration::new(4437052 * 365 * 24 * 60 * 60, 0);
         let colors = Colors::new(ThemeOption::Default);
         let date = Date::from(end_time);